Move utils files to their folders
