+++ src/email/sendEmail/sendEmail.test.ts
@@ -2,9 +2,9 @@ import { getRandomEmailOptions } from "../../factories/emailOptionsFactory.js";
 import sendEmail from "./sendEmail.js";
 import transporter from "../transporter.js";
 import type SMTPTransport from "nodemailer/lib/smtp-transport/index.js";
-import smtpStatusCodes from "../../utils/smtpStatusCodes.js";
 import type { EmailOptions } from "../types.js";
 import { environment } from "../../loadEnvironments.js";
+import smtpStatusCodes from "../../constants/statusCodes/smtpStatusCodes.js";
 
 const {
   smtp: { emailSender },
