+++ src/openapi/index.ts
@@ -1,11 +1,10 @@
 import path from "path";
 import { fileURLToPath } from "url";
-import { loadJson } from "../utils/loadJson.js";
+import { loadJson } from "../utils/loadJson/loadJson";
 
 const __dirname = fileURLToPath(path.dirname(import.meta.url));
 const openApiConfigFile = path.join(__dirname, "openapi.json");
 
-console.log(openApiConfigFile);
 const openApiDocument = loadJson<Record<string, unknown>>(openApiConfigFile);
 
 export default openApiDocument;
