+++ src/server/routers/usersRouter/usersRouter.test.ts
@@ -5,7 +5,7 @@ import { MongoMemoryServer } from "mongodb-memory-server";
 import connectDatabase from "../../../database/connectDatabase";
 import mongoose from "mongoose";
 import app from "../../app.js";
-import httpStatusCodes from "../../../utils/httpStatusCodes.js";
+import httpStatusCodes from "../../../constants/statusCodes/httpStatusCodes.js";
 import User from "../../../database/models/User.js";
 import type {
   CustomTokenPayload,
