--- src/utils/httpStatusCodes.ts
@@ -1,19 +1,0 @@
-const httpStatusCodes = {
-  successCodes: {
-    okCode: 200,
-    createdCode: 201,
-  },
-
-  clientErrors: {
-    badRequestCode: 400,
-    unauthorizedCode: 401,
-    notFoundCode: 404,
-    conflictCode: 409,
-  },
-
-  serverErrors: {
-    internalServerErrorCode: 500,
-  },
-};
-
-export default httpStatusCodes;
