--- src/utils/smtpStatusCodes.ts
@@ -1,7 +1,0 @@
-const smtpStatusCodes = {
-  positiveCompletion: {
-    okCode: 250,
-  },
-};
-
-export default smtpStatusCodes;
